halo2_poseidon

Trait Spec

pub trait Spec<F: Field, const T: usize, const RATE: usize>: Debug {
    // Required methods
    fn full_rounds() -> usize;
    fn partial_rounds() -> usize;
    fn sbox(val: F) -> F;
    fn secure_mds() -> usize;
    fn constants() -> (Vec<[F; T]>, Mds<F, T>, Mds<F, T>);
}

A specification for a Poseidon permutation.

Required Methods

fn full_rounds() -> usize

The number of full rounds for this specification.

This must be an even number.

fn partial_rounds() -> usize

The number of partial rounds for this specification.

fn sbox(val: F) -> F

The S-box for this specification.

fn secure_mds() -> usize

Side-loaded index of the first correct and secure MDS that will be generated by the reference implementation.

This is used by the default implementation of Spec::constants. If you are hard-coding the constants, you may leave this unimplemented.

fn constants() -> (Vec<[F; T]>, Mds<F, T>, Mds<F, T>)

Generates (round_constants, mds, mds^-1) corresponding to this specification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.